+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Shortcode ```[alert]```
 - Shortcode ```[badge]```
 - Shortcode ```[breadcrumb]``` and ```[breadcrumb-item]```
+- Shortcode ```[button]``` and ```[button-group]``` and ```[button-toolbar]```